Add Voting Banner, Fix Small Bugs and CSS Tweaks
Add a telos branded banner to the ballots-list page to improve page appearance.
Fix a small bug that caused a user navigating to a ballot via direct link to have the modal flash open and closed.
Add a max-width of 1200px to the ballot view modal to limit horizontal space on very large resolution screens.
Add a "no pdf found" icon to ballots that don't have an IPFS attachment.